mexico region codes
Mexico region codes are now a single tibble. This is adjusted for in the Mexico cleaning code. Level 2 Mexico codes are already downloaded in the data, and the Level 1 ISO codes are joined on based on that (as level 2 codes are an expansion of a different type of level 1 code, also stored in the level 1 tibble).